refactor mg doc and rw docs
